add , update expenses
