+++ app/src/main/java/com/example/dramebaz/shg/Presenter.java
@@ -102,59 +102,4 @@ public class Presenter {
         return !TextUtils.isEmpty(target) && android.util.Patterns.EMAIL_ADDRESS.matcher(target).matches();
     }
 
-    public static Map<String,Integer> getUsersShareMap(Context context, final Float cost, final String type, final Integer id){
-        final Map userShareMap = new LinkedHashMap<>();
-        // get current user id
-        SharedPreferences pref =
-                PreferenceManager.getDefaultSharedPreferences(context);
-        final int currentUserId =  pref.getInt("currentUserId", 0);
-        // check type
-        if(type.equals("group")){
-            // get all members id in case of group
-            SplitwiseRestClient client = RestApplication.getSplitwiseRestClient();
-            client.getGroup(new JsonHttpResponseHandler() {
-                @Override
-                public void onSuccess(int statusCode, Header[] headers, JSONObject json) {
-                    try {
-                        // divide cost between all members
-                        // paid share will be of current user id
-                        JSONObject group = json.getJSONObject("group");
-                        JSONArray members = group.getJSONArray("members");
-                        int length = members.length();
-                        for(int i=0;i<length;i++){
-                            JSONObject member = members.getJSONObject(i);
-                            if(member.getInt("id")!= currentUserId){
-                                userShareMap.put("users__"+i+"__user_id", member.getInt("id"));
-                                userShareMap.put("users__"+i+"__paid_share", 0);
-                                userShareMap.put("users__"+i+"__owed_share", cost/length);
-                            }else {
-                                userShareMap.put("users__"+i+"__user_id", member.getInt("id"));
-                                userShareMap.put("users__"+i+"__paid_share", cost);
-                                userShareMap.put("users__"+i+"__owed_share", cost/length);
-                            }
-                        }
-                        Log.i("get_group", json.toString());
-
-                    } catch (Exception e) {
-                        Log.e("FAILED get_group", "json_parsing", e);
-                    }
-                }
-
-                @Override
-                public void onFailure(int statusCode, Header[] headers, String responseString, Throwable throwable) {
-                    super.onFailure(statusCode, headers, responseString, throwable);
-                }
-            }, id);
-
-        }else{
-            userShareMap.put("users__0__user_id", currentUserId);
-            userShareMap.put("users__0__paid_share", cost);
-            userShareMap.put("users__0__owed_share", cost/2);
-            userShareMap.put("users__1__user_id", id);
-            userShareMap.put("users__1__paid_share", 0);
-            userShareMap.put("users__1__owed_share", cost/2);
-        }
-
-        return userShareMap;
-    }
 }
